Extract components from RecoveryView
